correc. ft_atoi

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: gsousa-l <gsousa-l@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/16 15:12:44 by gsousa-l          #+#    #+#             */
-/*   Updated: 2021/02/20 18:22:47 by gsousa-l         ###   ########.fr       */
+/*   Updated: 2021/02/20 20:27:26 by gsousa-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_atoi(const char *nptr)
 {
-	long	num;
+	int	num;
 	int		count;
 	int		signal;
